Chat-Bot/requirements.txt
flask
python-dotenv
google-generativeai
gunicorn
requests
